simplify synchronous dummy transmission module
